Media Clock Recovery implemented
Support for IEEE 1722 Clock Reference Format (CRF) streams
is added. CRF streams can be output based on the rate that audio
is consumed. CRF streams can be received from an external source
and used to synchronize an output audio stream. The output
audio stream can be synchronized to an input audio stream.

Change-Id: I6b6cc1df64ac48c42c948f1a7b2f4e1fd0c3f278
CRs-Fixed: 1019444